Add disqus code snippet
We'll use disqus to manage the commenting functionality on the site. Disqus
provides a code snippet in <script> tags that will connect back to the disqus
service. This snippet is added to the post template (in _layouts) inside an
if liquid tag, so that comments can be enabled by including
comments: true
in the YAML front matter for the post.