remove 5 post limit mistake on blog page
In the long term, still need to develop pagination for the blog page
once the number of posts increases. For now, the blog page will just
load all of the posts, and the home page will still be limited to 5
posts and then have a link directing users to the blog page for more.